Delete old classes
