+++ src/main/java/com/palantir/gradle/gitversion/VersionDetails.java
@@ -1,130 +1,133 @@
-/*
- * Copyright 2016 Palantir Technologies
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * <http://www.apache.org/licenses/LICENSE-2.0>
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package com.palantir.gradle.gitversion;
 
-import java.io.Serializable;
+import org.eclipse.jgit.api.Git;
+import org.eclipse.jgit.api.errors.GitAPIException;
+import org.eclipse.jgit.lib.Constants;
+import org.eclipse.jgit.lib.ObjectId;
+import org.eclipse.jgit.lib.Ref;
+
+import java.io.IOException;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class VersionDetails implements Serializable, IVersionDetails {
+public class VersionDetails {
 
-    // Gradle returns 'unspecified' when no version is set
-    private static final String UNSPECIFIED_VERSION = "unspecified";
-    private static final long serialVersionUID = -7340444937169877612L;
+    private static final int VERSION_ABBR_LENGTH = 10;
 
-    private final String description;
-    private final String gitHash;
-    private final String gitHashFull;
-    private final String branchName;
-    private final boolean isClean;
+    private final Git git;
+    private final GitVersionArgs args;
 
-    public VersionDetails(String description, String gitHash, String gitHashFull, String branchName, boolean isClean) {
-        this.description = description;
-        this.gitHash = gitHash;
-        this.gitHashFull = gitHashFull;
-        this.branchName = branchName;
-        this.isClean = isClean;
+    private volatile String maybeCachedDescription = null;
+
+    public VersionDetails(Git git, GitVersionArgs args) {
+        this.git = git;
+        this.args = args;
     }
 
-    @Override
-    public String getVersion() {
-        if (description == null) {
-            return UNSPECIFIED_VERSION;
+    private String description() {
+        if (maybeCachedDescription == null) {
+            String rawDescription = expensiveComputeRawDescription();
+            this.maybeCachedDescription = rawDescription == null ?
+                    rawDescription : rawDescription.replaceFirst("^" + args.getPrefix(), "");
         }
 
-        return description + (isClean ? "" : ".dirty");
+        return maybeCachedDescription;
     }
 
-    @Override
+    private String expensiveComputeRawDescription() {
+        String nativeGitDescribe = new NativeGitDescribe(git.getRepository().getDirectory(), git)
+                .describe(args.getPrefix());
+        String jgitDescribe = new JGitDescribe(git)
+                .describe(args.getPrefix());
+
+        // If native failed, return JGit one
+        if (nativeGitDescribe == null) {
+            return jgitDescribe;
+        }
+
+        // If native succeeded, make sure it's same as JGit one
+        if (!nativeGitDescribe.equals(jgitDescribe)) {
+            throw new IllegalStateException(String.format("Inconsistent git describe: native was %s and jgit was %s. " +
+                    "Please report this on github.com/palantir/gradle-git-version", nativeGitDescribe, jgitDescribe));
+        }
+
+        return jgitDescribe;
+    }
+
+    public String getVersion() {
+        if (description() == null) {
+            return "unspecified";
+        }
+
+        return description() + (isClean() ? "" : ".dirty");
+    }
+
     public boolean getIsCleanTag() {
-        return isClean && descriptionIsPlainTag();
+        return isClean() && descriptionIsPlainTag();
     }
 
-    @Override
+    private boolean descriptionIsPlainTag() {
+        return !Pattern.matches(".*g.?[0-9a-fA-F]{3,}", description());
+    }
+
     public int getCommitDistance() {
         if (descriptionIsPlainTag()) {
             return 0;
         }
 
-        Matcher match = Pattern.compile("(.*)-([0-9]+)-g.?[0-9a-fA-F]{3,}").matcher(description);
+        Matcher match = Pattern.compile("(.*)-([0-9]+)-g.?[0-9a-fA-F]{3,}").matcher(description());
         return match.matches() ? Integer.valueOf(match.group(2)) : null;
     }
 
-    private boolean descriptionIsPlainTag() {
-        return !Pattern.matches(".*g.?[0-9a-fA-F]{3,}", description);
-    }
-
-    @Override
     public String getLastTag() {
         if (descriptionIsPlainTag()) {
-            return description;
+            return description();
         }
 
-        Matcher match = Pattern.compile("(.*)-([0-9]+)-g.?[0-9a-fA-F]{3,}").matcher(description);
+        Matcher match = Pattern.compile("(.*)-([0-9]+)-g.?[0-9a-fA-F]{3,}").matcher(description());
         return match.matches() ? match.group(1) : null;
     }
 
-    @Override
+    private boolean isClean() {
+        try {
+            return git.status().call().isClean();
+        } catch (GitAPIException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
     public String getGitHash() {
-        return gitHash;
+        String gitHashFull = getGitHashFull();
+        if (gitHashFull == null) {
+            return null;
+        }
+
+        return gitHashFull.substring(0, VERSION_ABBR_LENGTH);
     }
 
-    @Override
     public String getGitHashFull() {
-        return gitHashFull;
+        try {
+            ObjectId objectId = git.getRepository().findRef(Constants.HEAD).getObjectId();
+            if (objectId == null) {
+                return null;
+            }
+
+            return objectId.name();
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 
-    /** @return null if the repository in detached HEAD mode */
-    @Override
     public String getBranchName() {
-        return branchName;
-    }
+        try {
+            Ref ref = git.getRepository().findRef(git.getRepository().getBranch());
+            if (ref == null) {
+                return null;
+            }
 
-    @Override
-    public String toString() {
-        return "VersionDetails{" +
-                "description='" + description + '\'' +
-                ", gitHash='" + gitHash + '\'' +
-                ", gitHashFull='" + gitHashFull + '\'' +
-                ", branchName='" + branchName + '\'' +
-                ", isClean=" + isClean +
-                '}';
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        VersionDetails that = (VersionDetails) o;
-
-        if (isClean != that.isClean) return false;
-        if (description != null ? !description.equals(that.description) : that.description != null) return false;
-        if (gitHash != null ? !gitHash.equals(that.gitHash) : that.gitHash != null) return false;
-        if (gitHashFull != null ? !gitHashFull.equals(that.gitHashFull) : that.gitHashFull != null) return false;
-        return branchName != null ? branchName.equals(that.branchName) : that.branchName == null;
-    }
-
-    @Override
-    public int hashCode() {
-        int result = description != null ? description.hashCode() : 0;
-        result = 31 * result + (gitHash != null ? gitHash.hashCode() : 0);
-        result = 31 * result + (gitHashFull != null ? gitHashFull.hashCode() : 0);
-        result = 31 * result + (branchName != null ? branchName.hashCode() : 0);
-        result = 31 * result + (isClean ? 1 : 0);
-        return result;
+            return ref.getName().substring(Constants.R_HEADS.length());
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 }
